+++ AdsLib/AdsDevice.h
@@ -81,15 +81,15 @@ struct AdsDevice {
     uint32_t GetTimeout() const;
     void SetTimeout(const uint32_t timeout) const;
 
-    long ReadReqEx2(uint32_t group, uint32_t offset, uint32_t length, void* buffer, uint32_t* bytesRead) const;
+    long ReadReqEx2(uint32_t group, uint32_t offset, size_t length, void* buffer, uint32_t* bytesRead) const;
     long ReadWriteReqEx2(uint32_t    indexGroup,
                          uint32_t    indexOffset,
-                         uint32_t    readLength,
+                         size_t      readLength,
                          void*       readData,
-                         uint32_t    writeLength,
+                         size_t    writeLength,
                          const void* writeData,
                          uint32_t*   bytesRead) const;
-    long WriteReqEx(uint32_t group, uint32_t offset, uint32_t length, const void* buffer) const;
+    long WriteReqEx(uint32_t group, uint32_t offset, size_t length, const void* buffer) const;
 
     AdsResource<const AmsNetId> m_NetId;
     const AmsAddr m_Addr;
